requirements.txt
flask==3.0.3
flask-jwt-extended==4.6.0
flask-cors==4.0.1
bcrypt==4.1.3
python-dotenv==1.0.1
fpdf2==2.7.9
google-auth-oauthlib==1.2.0
google-api-python-client==2.131.0
